Add parameters for scanning #5
