--- server/models/facets.js
@@ -1,3 +1,0 @@
-// our example model is just an Array
-var facets = [];
-export default facets;
